+++ Inspinia_MVC5_SeedProject/Models/cNotaCredito.cs
@@ -30,7 +30,7 @@ namespace ERP_GMEDINA.Models
 
         [Display(Name = "Sucursal")]
         [Required(AllowEmptyStrings = false, ErrorMessage = "*El campo {0} es requerido")]
-        public short suc_Id { get; set; }
+        public int suc_Id { get; set; }
 
         [Display(Name = "Anulada")]
         [Required(AllowEmptyStrings = false, ErrorMessage = "*El campo {0} es requerido")]
